refactor(train): split stack.py into planning + base package
Pure structural split of unirl/train/stack.py (now ~870 lines) along its one real
seam: the micro-batch PLANNING (Plan types, token-budget bin-packing math, the
MicroPlanner / CountPlanner / TokenBudgetPlanner strategies) vs the distributed
DRIVER (TrainStack). They have different dependency and test profiles —

  - planning.py: pure python, no FSDP/Ray, CPU-unit-tested (test_packing)
  - base.py: the FSDP/Ray Remote training driver

No behaviour change. The planner classes stay WITH the packing functions they wrap
in planning.py (they are bound — a planner is just the OO face of the functions);
the split only separates that planning unit from the driver. Public surface
unchanged: __init__ re-exports TrainStack / TrainStepResult / the planners /
_build_micro_batch_slices, so recipes (`_target_: unirl.train.stack.TrainStack` /
`...TokenBudgetPlanner`) and `from unirl.train.stack import ...` all resolve as
before. test_packing imports the private helpers from the `planning` submodule.

Verified on Qwen env: ruff clean, 1004 _target_ paths resolve, test_packing passes.
